verify_commands_parameters function
Make a function that verifies that the parameters inside the command list are valid parameters. Also fixed errors along the way while making it, such as making a new list to store the valid commands. Also added module doc strings and annotations within the functions to make it more readable. Finally made a unittest to test the new function.