New Module: runcommands.py + sort_set_up_list function + unittest
Made a new module that is in charge of running commands, but in the process realized the need to further sort the set up list. This new function moves PROPAGATE commands to the end of the list so that the length and devices are established beforehand. Also made a unittest for this function to ensure coverage.