sort_command_list + cancel attribute
Changed sort command so that the CANCEL commands are run first. Also made a cancel_received attribute and a cancel_time attribute in the device class so that I can use them as conditionals so that the simulation output will only run based on certain conditions I could not have met without the new attributes.