Annotations (all modules) and Testing edge cases
Made annotations on mostly all created modules and functions, with the functions that are not tested, I wrote an explanation above the function explaining why it was not tested. Also testing edge cases with running the function, making sure it works in all cases. Specifically, added an if statement to only recursion in propagation_cancel if the propagate commands list is more than length 0.